requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0


dj-database-url==2.1.0
# Pillow==10.1.0
Pillow==11.0.0
# psycopg2-binary==2.9.9
psycopg[binary]==3.2.9

# UI and Forms
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# File Handling
django-cleanup==8.0.0

# Development Tools
django-extensions==3.2.3

# Database and Caching
redis==5.0.1

# Background Tasks
celery==5.3.4

# Data Processing and Utilities
Faker==19.12.0
python-slugify==8.0.1
python-dateutil==2.8.2

# HTTP Requests
requests==2.31.0

# WhatsApp/API Integration
twilio==8.10.0

# Monitoring and Logging
sentry-sdk[django]==1.38.0

# Testing
pytest-django==4.7.0
factory-boy==3.3.0
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0